Flask-Pytorch/requirements.txt
torch==1.12.0
torchvision==0.13.0
facenet-pytorch==2.5.3
pillow==9.5.0
numpy==1.21.6
flask==2.2.5
opencv-python==4.5.5.64